third_party/blink/web_tests/platform/win11-arm64/http/tests/misc/resource-timing-sizes-content-encoding-expected.txt

This is a testharness.js-based test.
[FAIL] PerformanceResourceTiming sizes compressed content test
  assert_between_exclusive: encodedBodySize expected a number greater than 0 and less than 1250 but got 1986
Harness: the test ran to completion.
